ended firs option

+++ AppGestionarAnimales/accesoDatos/Datos.cs
@@ -56,7 +56,7 @@ namespace AppTiendaMascotas.accesoDatos
             return ds;
         }
 
-        public int ConsultarIngXEmpleado(int codEmpleado, DateTime fechaInicio, DateTime fechaFin)
+        public int ConsultarIngXEmpleado(long codEmpleado, DateTime fechaInicio, DateTime fechaFin)
         {
             int total = 0;
             OracleConnection connection = new OracleConnection(cadenaConexion);
@@ -69,7 +69,7 @@ namespace AppTiendaMascotas.accesoDatos
                 command.CommandText = "paq_gerente.total_ingresos_empleado";
                 command.CommandType = CommandType.StoredProcedure;
 
-                command.Parameters.Add("p_codEmpleado", OracleDbType.Int32).Value = codEmpleado;
+                command.Parameters.Add("p_codEmpleado", OracleDbType.Decimal).Value = codEmpleado;
                 command.Parameters.Add("p_fechaInicio", OracleDbType.Date).Value = fechaInicio;
                 command.Parameters.Add("p_fechaFin", OracleDbType.Date).Value = fechaFin;
